+++ solver/src/main/java/org/chocosolver/solver/constraints/nary/flow/PropMinCostMaxFlow.java
@@ -1,7 +1,7 @@
 /*
  * This file is part of choco-solver, http://choco-solver.org/
  *
- * Copyright (c) 2023, IMT Atlantique. All rights reserved.
+ * Copyright (c) 2024, IMT Atlantique. All rights reserved.
  *
  * Licensed under the BSD 4-clause license.
  *
@@ -70,80 +70,33 @@ public class PropMinCostMaxFlow extends Propagator<IntVar> {
         int maxcost = this.cost.getUB();
         try {
             for (int i = toCheck.nextSetBit(0); i > -1; i = toCheck.nextSetBit(i + 1)) {
-                updateFlow4(i, maxcost);
+                updateFlow(i, maxcost);
             }
         } finally {
             toCheck.clear();
         }
     }
 
-    private void updateFlow0(int i, int minCost, int maxcost) throws ContradictionException {
+    private void updateFlow(int i, int maxcost) throws ContradictionException {
         int l = flows[i].getLB();
         int u = flows[i].getUB();
         int m = u;
         g.refresh(i, m);
         int c = minCostFlow(i, m);
         if (c > maxcost || c == -1) {
+            // u is the first value that does not work
+            u--;
             while (l <= u) {
                 m = (l + u) >>> 1;
                 g.refresh(i, m);
                 c = minCostFlow(i, m);
-                if (c == -1) {
-                    u = m - 1;
-                } else if (c > maxcost) {
-                    u = m - 1;
-                } else if (c < maxcost) {
-                    l = m + 1;
-                } else break;
-            }
-            flows[i].updateUpperBound(m, this);
-        }
-    }
-
-    private void updateFlow1(int i, int minCost, int maxcost) throws ContradictionException {
-        int c;
-        int l;
-        do {
-            l = flows[i].getLB();
-            g.refresh(i, l);
-            c = minCostFlow(i, l);
-        } while ((c == -1 || c < minCost) && flows[i].updateLowerBound(l + 1, this));
-        int u;
-        do {
-            u = flows[i].getUB();
-            g.refresh(i, u);
-            c = minCostFlow(i, u);
-        } while ((c == -1 || c > maxcost) && flows[i].updateUpperBound(u - 1, this));
-    }
-
-    private void updateFlow4(int i, int maxcost) throws ContradictionException {
-        int l = flows[i].getLB();
-        int u = flows[i].getUB();
-        int m = u;
-        g.refresh(i, m);
-        int c = minCostFlow(i, m);
-        if (c > maxcost || c == -1) {
-            while (l <= u) {
-                m = (l + u) >>> 1;
-                g.refresh(i, m);
-                c = minCostFlow(i, m);
-                if (c == -1) {
-                    u = m - 1;
-                } else if (c > maxcost) {
-                    u = m - 1;
-                } else if (c < maxcost) {
-                    l = m + 1;
-                } else break;
-            }
-            flows[i].updateUpperBound(m, this);
-            if (l == u + 1) {
-                u = flows[i].getUB();
-                g.refresh(i, u);
-                c = minCostFlow(i, u);
                 if (c == -1 || c > maxcost) {
-                    flows[i].updateUpperBound(u - 1, this);
+                    u = m - 1;
+                } else {// if (c <= maxcost) {
+                    l = m + 1;
                 }
             }
+            flows[i].updateUpperBound(u, this);
         }
     }
 
@@ -193,7 +146,6 @@ public class PropMinCostMaxFlow extends Propagator<IntVar> {
         final int[] d; // distance
         final int[] p; // predecessor
         final int[] b; // balance or demand on each vertex
-        final BitSet inqueue;
         final IntCircularQueue queue;
         int D;
 
@@ -204,7 +156,6 @@ public class PropMinCostMaxFlow extends Propagator<IntVar> {
             this.n2 = n - 2;
             this.d = new int[n];
             this.p = new int[n];
-            this.inqueue = new BitSet(n);
             this.queue = new IntCircularQueue(n);
             this.adj = new ArrayList[n];
             this.b = new int[n];
@@ -289,22 +240,17 @@ public class PropMinCostMaxFlow extends Propagator<IntVar> {
         Arrays.fill(g.d, Integer.MAX_VALUE);
         Arrays.fill(g.p, -1);
         g.d[s] = 0;
-        g.inqueue.clear();
         g.queue.clear();
         g.queue.addLast(s);
         while (!g.queue.isEmpty()) {
             int u = g.queue.pollFirst();
-            g.inqueue.clear(u);
             for (int i = 0; i < g.adj[u].size(); i++) {
                 Edge e = g.adj[u].get(i);
                 int v = e.to;
                 if (e.capacity > 0 && g.d[v] > g.d[u] + e.cost) {
                     g.d[v] = g.d[u] + e.cost;
                     g.p[v] = e.id; // not the predecessor but the arc
-                    if (!g.inqueue.get(v)) {
-                        g.inqueue.set(v);
-                        g.queue.addLast(v);
-                    }
+                    g.queue.addLast(v);
                 }
             }
         }