+++ solver/src/main/java/org/chocosolver/solver/propagation/PropagationEngine.java
@@ -1,7 +1,7 @@
 /*
  * This file is part of choco-solver, http://choco-solver.org/
  *
- * Copyright (c) 2023, IMT Atlantique. All rights reserved.
+ * Copyright (c) 2024, IMT Atlantique. All rights reserved.
  *
  * Licensed under the BSD 4-clause license.
  *
@@ -76,6 +76,10 @@ public class PropagationEngine {
      */
     protected Propagator<?> lastProp;
     /**
+     * The last scheduled variable
+     */
+    protected Variable lastVar;
+    /**
      * One bit per queue: true if the queue is not empty.
      */
     private int notEmpty;
@@ -93,7 +97,7 @@ public class PropagationEngine {
      * propagation engine.
      * when set to '0b10', this works as a variable-oriented propagation engine.
      */
-    private final byte hybrid;
+    private byte hybrid;
     /**
      * For dynamic addition, avoid creating a new lambda at each call
      */
@@ -104,6 +108,11 @@ public class PropagationEngine {
             awake_queue.addLast(propagator);
         }
     };
+
+    /**
+     * A propagation insight to collect information about the propagation
+     */
+    private PropagationInsight insight = PropagationInsight.VOID;
 
     /**
      * A seven-queue propagation engine.
@@ -126,6 +135,8 @@ public class PropagationEngine {
         this.awake_queue = new CircularQueue<>(16);
         this.dynPropagators = new DynPropagators();
         this.propagators = new ArrayList<>();
+        //0b00: cstr-ori
+        //0b10: var-ori
         this.hybrid = model.getSettings().enableHybridizationOfPropagationEngine();
         this.sat = sat;
     }
@@ -160,14 +171,14 @@ public class PropagationEngine {
             }
             if (model.getSettings().sortPropagatorActivationWRTPriority()) {
                 propagators.sort(
-                    (p1, p2) -> {
-                        int p = p1.getPriority().getValue() - p2.getPriority().getValue();
-                        if (p == 0) {
-                            return p1.getNbVars() - p2.getNbVars();
-                        } else {
-                            return p;
-                        }
-                    });
+                        (p1, p2) -> {
+                            int p = p1.getPriority().getValue() - p2.getPriority().getValue();
+                            if (p == 0) {
+                                return p1.getNbVars() - p2.getNbVars();
+                            } else {
+                                return p;
+                            }
+                        });
             }
             for (int i = 0; i < propagators.size(); i++) {
                 Propagator<?> propagator = getPropagator(i);
@@ -212,19 +223,27 @@ public class PropagationEngine {
      */
     public void propagate() throws ContradictionException {
         propagateSat();
+        insight.clear();
         activatePropagators();
         do {
             manageModifications();
             for (int i = nextNotEmpty(); i > -1; i = nextNotEmpty()) {
                 assert !pro_queue[i].isEmpty() : "try to pop a propagator from an empty queue";
                 lastProp = pro_queue[i].pollFirst();
+                insight.cardinality(lastProp);
                 if (pro_queue[i].isEmpty()) {
                     notEmpty &= ~(1 << i);
                 }
                 // revision of the variable
                 lastProp.unschedule();
                 delayedPropagationType = 0;
-                propagateEvents();
+                try {
+                    propagateEvents();
+                    insight.update(lastProp, lastVar, false);
+                } catch (ContradictionException cex) {
+                    insight.update(lastProp, lastVar, true);
+                    throw cex;
+                }
                 if (hybrid < 0b01) {
                     manageModifications();
                 }
@@ -256,6 +275,7 @@ public class PropagationEngine {
 
     /**
      * Checks if some propagators were added or have to be propagated on backtrack
+     *
      * @throws ContradictionException if a propagation fails
      */
     private void activatePropagators() throws ContradictionException {
@@ -317,6 +337,7 @@ public class PropagationEngine {
             notEmpty = notEmpty & ~(1 << i);
         }
         lastProp = null;
+        lastVar = null;
     }
 
     /**
@@ -336,6 +357,7 @@ public class PropagationEngine {
             }
             assert found : variable + " not in scope of " + cause;
         }
+        insight.modifiy(variable);
         if (!variable.isScheduled()) {
             var_queue.addLast(variable);
             variable.schedule();
@@ -363,7 +385,7 @@ public class PropagationEngine {
     /**
      * @return the propagator Event Type's mask for delayed propagation
      */
-    int getDelayedPropagation(){
+    int getDelayedPropagation() {
         return delayedPropagationType;
     }
 
@@ -387,6 +409,14 @@ public class PropagationEngine {
         }
     }
 
+    public void setInsight(PropagationInsight insight) {
+        this.insight = insight;
+    }
+
+    public void setHybrid(byte hybrid) {
+        this.hybrid = hybrid;
+    }
+
     /**
      * Reset the propagation engine.
      */
@@ -405,6 +435,7 @@ public class PropagationEngine {
         notEmpty = 0;
         init = false;
         lastProp = null;
+        lastVar = null;
     }
 
     public void ignoreModifications() {