+++ solver/src/test/java/org/chocosolver/solver/search/BlackBoxTest.java
@@ -1,7 +1,7 @@
 /*
  * This file is part of choco-solver, http://choco-solver.org/
  *
- * Copyright (c) 2023, IMT Atlantique. All rights reserved.
+ * Copyright (c) 2024, IMT Atlantique. All rights reserved.
  *
  * Licensed under the BSD 4-clause license.
  *
@@ -41,6 +41,8 @@ public class BlackBoxTest {
                 {(Function<IntVar[], AbstractStrategy<IntVar>>) Search::domOverWDegRefSearch},
                 {(Function<IntVar[], AbstractStrategy<IntVar>>) Search::failureRateBasedSearch},
                 {(Function<IntVar[], AbstractStrategy<IntVar>>) Search::failureLengthBasedSearch},
+                {(Function<IntVar[], AbstractStrategy<IntVar>>) Search::pickOnDom},
+                {(Function<IntVar[], AbstractStrategy<IntVar>>) Search::pickOnFil},
         };
     }